% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{Total abundance (number of individuals) of each prey species in a stomac content dataset}
\usage{
abundance(data_stomac, min_abundance, quantile_abundance, proportion_abundance)
}
\arguments{
\item{data_stomac}{Raw data stomach content with number of a given prey (ScientificName_accepted) in all the tractus (Nbr.total.tractus) for related tag fish (N.poissons).}

\item{min_abundance}{Minimal number of individual found in all tractus to consider the species}

\item{quantile_abundance}{Minimal quantil in all tractus to consider the species}

\item{proportion_abundance}{Minimal proportion of individual found in all tractus to consider the species}
}
\value{
dataframe Prey species, Total number of preys
}
\description{
Total abundance (number of individuals) of each prey species in a stomac content dataset
}
\examples{
library(chopin.project)
abundance(data_stomac = stomac_soles, min_abundance = 5, quantile_abundance = 0.95, proportion_abundance = 0.95)
}
