% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distri_diet.R
\name{distri_diet}
\alias{distri_diet}
\title{distri_diet Title}
\usage{
distri_diet(data_stomac)
}
\arguments{
\item{data_stomac}{Raw data stomach content with number of a given prey (ScientificName_accepted) in all the tractus (Nbr.total.tractus) for related tag fish (N.poissons)}
}
\value{
Matrix of proportion of each prey compared to the total number of preys in each fish tag
}
\description{
distri_diet Title
}
\examples{
library(chopin.project)
distri_diet(stomac_soles)
}
