% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_res_diet.R
\name{get_res_diet}
\alias{get_res_diet}
\title{Pie of stomach content with proportions of each major prey}
\usage{
get_res_diet(main, grp, min_occur, min_abond)
}
\arguments{
\item{main}{Legend for the group of soles studied}

\item{grp}{Raw data stomach content with number of a given prey (ScientificName_accepted) in all the tractus (Nbr.total.tractus) for related tag fish (N.poissons)}

\item{min_occur}{Minimal occurence of preys in tractus considered in the plot (0-1)}

\item{min_abond}{Minimal abondance of preys in tractus considered in the plot (0-1)}
}
\value{
Pie graph
}
\description{
Description
}
\examples{
library(chopin.project)
res_all_soles = get_res_diet(main = "ALL SOLES", grp=stomac_soles, min_abond =0.1, min_occur = 0.1)
}
