% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_group_zone_saison.R
\name{GRAPH_group_zone_saison}
\alias{GRAPH_group_zone_saison}
\title{graph_group_zone_saison  FONCTION GRAPHIQUE TAXON, ZONE, SAISON}
\usage{
GRAPH_group_zone_saison(data, label_data, variable, label_y, wd)
}
\arguments{
\item{data}{}

\item{label_data}{}

\item{variable}{}

\item{label_y}{}

\item{wd}{}
}
\value{
1
}
\description{
graph_group_zone_saison  FONCTION GRAPHIQUE TAXON, ZONE, SAISON
}
\examples{
sub_benthos_contam <- benthos_contam |> 
  select(grp, season, zone, starts_with("somme")  )
sub_benthos_contam <- as.data.frame(sub_benthos_contam)

GRAPH_group_zone_saison(data = sub_benthos_contam, label_data = "benthos", variable = "sommePCB_ng_gdw", label_y=expression(paste(sum(),"PCB")), wd = "")
}
