% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_grp.R
\name{graph_grp}
\alias{graph_grp}
\title{Dotplot}
\usage{
graph_grp(data, grp, variable, label_data, label_y, title)
}
\arguments{
\item{data}{}

\item{variable}{}

\item{label_data}{}

\item{label_y}{}

\item{title}{}

\item{group}{}

\item{wd}{}
}
\description{
Description
}
\examples{
sub_benthos_contam <- benthos_contam |> 
  select(grp, season, zone, starts_with("somme")  )
sub_benthos_contam <- as.data.frame(sub_benthos_contam)

graph_grp(
  data = sub_benthos_contam,
  grp = "grp",
  variable = "sommePCB_ng_gdw",
  label_data = "soles et benthos",
  label_y = expression(paste(sum(), "PCB")),
  title = "GROUPE"
)
}
