% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_grp.R
\name{list_grp}
\alias{list_grp}
\title{Create subdatasets of stomach content data depending on sole age and sector.}
\usage{
list_grp(dataset)
}
\arguments{
\item{dataset}{Raw data stomach content with number of a given prey (ScientificName_accepted) in all the tractus (Nbr.total.tractus) for related tag fish (N.poissons). Studied columns are 'espèce.stade' and 'secteur'.}
}
\value{
a list of stomach content subdatasets of interest: G0_CH, G0_EM, G0_FN, G0_FS, G1_CH, G1_EM, G1_FN, G1_FS, G2_FN, G2_EM
}
\description{
Description
}
\examples{
library(chopin.project)
list_grps = list_grp(stomac_soles)
}
