% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.R
\name{occ}
\alias{occ}
\title{occ Fonction de calcul de la fréquence d'occurence (concentration != 0) parmi les échantillons analysés (non NA)}
\usage{
occ(vector)
}
\arguments{
\item{vector}{Vecteur contenant des valeurs de contamination ou des valeurs manquantes}
}
\value{
value Proportion de valeurs non nulles parmis
}
\description{
occ Fonction de calcul de la fréquence d'occurence (concentration != 0) parmi les échantillons analysés (non NA)
}
\examples{
 vec <- c(0, 1, 5, NA)
 occ(vector = vec)
}
