% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence.R
\name{occurence}
\alias{occurence}
\title{Occurrence (%number over number of all preys) of each prey species in a stomac content dataset}
\usage{
occurence(data_stomac)
}
\arguments{
\item{data_stomac}{Raw data stomach content with number of a given prey (ScientificName_accepted) in all the tractus (Nbr.total.tractus) for related tag fish (N.poissons).}
}
\value{
dataframe Prey species, Total proportion in all tractus
}
\description{
Occurrence (%number over number of all preys) of each prey species in a stomac content dataset
}
\examples{
library(chopin.project)
occurence(stomac_soles)
}
