% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart_diet.R
\name{piechart_diet}
\alias{piechart_diet}
\title{Piechart of diet for a group of soles}
\usage{
piechart_diet(res_pie)
}
\description{
with type of color by taxon
}
\examples{
library(chopin.project)
library(tidyverse)
library(scales)

res_G0_emb_juin = get_res_diet(
  main = "G0 Embouchure printemps",
  grp = stomac_soles[which(
    stomac_soles$secteur == "Embouchure" &
      stomac_soles$espèce.stade == "Sole G0" &
      stomac_soles$Campagne == "Print-17"
  ), ],
  min_abond = 0.1,  min_occur = 0.1
)
piechart_diet(res_pie = res_G0_emb_juin$`med \%abun`)
}
