% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_somme_facteur.R
\name{stat_somme_facteur}
\alias{stat_somme_facteur}
\title{stat_somme_facteur Effectifs analysés et sommes par famille de contaminant en fonction d'un facteur}
\usage{
stat_somme_facteur(dataset, facteur, title)
}
\value{
dataframe
}
\description{
stat_somme_facteur Effectifs analysés et sommes par famille de contaminant en fonction d'un facteur
}
\examples{
data("benthos_contam")
stat_somme_facteur(dataset=benthos_contam, facteur = "grp", title="")
}
