% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_effect.R
\name{test_effect}
\alias{test_effect}
\title{test_effect Test l'effet d'une variable qualitative sur une variable quantitative}
\usage{
test_effect(dataset, variable, factor)
}
\arguments{
\item{dataset}{dataframe contenant les deux variables à étudier}

\item{variable}{vecteur du dataframe - variable quantitative}

\item{factor}{vecteur du dataframe - variable qualitative}
}
\value{
1
}
\description{
Lance les tests suivants :
- Kruskal-Wallis Rank Sum Test
- Pairwise Wilcoxon Rank Sum Tests
- Wilcoxon Rank Sum and Signed Rank Tests
}
\examples{
test_effect(dataset = iris, variable = "Sepal.Length", factor = "Species")
}
